把雜亂的javascript和css程式碼重複的部分切分到另一個檔案裡
