+++ static/pic/icon_delete.png
