+++ static/signinup.css
@@ -1,0 +1,158 @@
+#dark {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  display: none;
+  opacity: 0.4;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  z-index: 31;
+  filter: contrast(0) brightness(0.5);
+  background-color: white;
+}
+#preSignin {
+  display: block;
+}
+#afterSignin {
+  display: none;
+}
+#signin {
+  border-radius: 6px;
+  background: #ffffff;
+  box-shadow: 0px 4px 60px #aaaaaa;
+  position: fixed;
+  left: 39%;
+  z-index: 35;
+  width: 340px;
+  height: 275px;
+  top: 12%;
+  display: none;
+  animation-name: movedown;
+  animation-duration: 500ms;
+  animation-delay: 0s;
+  animation-fill-mode: both;
+  animation-timing-function: ease-out;
+}
+@keyframes movedown {
+  0% {
+    bottom: 12%;
+  }
+  25% {
+    bottom: 8%;
+  }
+  50% {
+    bottom: 6%;
+  }
+  100% {
+    bottom: 4%;
+  }
+}
+#signup {
+  border-radius: 6px;
+  background: #ffffff;
+  box-shadow: 0px 4px 60px #aaaaaa;
+  position: fixed;
+  left: 39%;
+  z-index: 35;
+  width: 340px;
+  height: 332px;
+  bottom: 49%;
+  display: none;
+}
+.decorator {
+  width: 340px;
+  height: 10px;
+  background: linear-gradient(270deg, #337788 0%, #66aabb 100%);
+  border-radius: 6px 6px 0px 0px;
+}
+.sign_title {
+  text-align: center;
+  width: 310px;
+  height: 27px;
+  font-weight: 700;
+  font-size: 24px;
+  line-height: 24px;
+  color: #666666;
+  margin: 15px;
+}
+.username {
+  width: 310px;
+  height: 47px;
+  border: 1px solid #cccccc;
+  box-sizing: border-box;
+  border-radius: 5px;
+  margin-left: 15px;
+  margin-right: 15px;
+  margin-bottom: 10px;
+}
+.email {
+  width: 310px;
+  height: 47px;
+  border: 1px solid #cccccc;
+  box-sizing: border-box;
+  border-radius: 5px;
+  margin-left: 15px;
+  margin-right: 15px;
+  margin-bottom: 10px;
+}
+.password {
+  width: 310px;
+  height: 47px;
+  border: 1px solid #cccccc;
+  box-sizing: border-box;
+  border-radius: 5px;
+  margin-left: 15px;
+  margin-right: 15px;
+  margin-bottom: 10px;
+}
+.submit {
+  width: 310px;
+  height: 47px;
+  background: #448899;
+  border-radius: 5px;
+  border: none;
+  margin-left: 15px;
+  margin-right: 15px;
+  margin-bottom: 10px;
+  color: #ffffff;
+}
+.submit:hover {
+  cursor: pointer;
+}
+.switch {
+  text-align: center;
+  width: 310px;
+  height: 22px;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 13px;
+  color: #666666;
+  border: none;
+  background-color: white;
+  margin-left: 15px;
+}
+.switch:hover {
+  cursor: pointer;
+}
+#up_message {
+  text-align: center;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 13px;
+  color: red;
+  margin-top: 5px;
+  margin-bottom: 9px;
+  display: none;
+}
+#in_message {
+  text-align: center;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 13px;
+  color: red;
+  margin-top: 5px;
+  margin-bottom: 9px;
+  display: none;
+}
