+++ templates/attraction.html
@@ -4,9 +4,11 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/png" sizes="32x32" href="../static/bx-building-house.svg">
+    <link rel="icon" type="image/png" sizes="32x32" href="../static/pic/bx-building-house.svg">
     <title>台北一日遊</title>
     <link href="../static/attraction.css" rel="stylesheet"/>
+    <link href="../static/same.css" rel="stylesheet"/>
+    <link href="../static/signinup.css" rel="stylesheet"/>
 </head>
 <body onload="getattraction()">
     <div class="title">
@@ -40,8 +42,8 @@
     <div id="dark" onclick="darkover()"></div>
     <div class="upper" id="upper">        
     <div id="pic">
-        <img src="../static/btn_leftArrow.png" id="pic3" onmouseover="enter1()" onmouseout="leave1()" onclick="switchto1()">
-        <img src="../static/btn_rightArrow.png" id="pic4" onmouseover="enter2()" onmouseout="leave2()" onclick="switchto2()">
+        <img src="../static/pic/btn_leftArrow.png" id="pic3" onmouseover="enter1()" onmouseout="leave1()" onclick="switchto1()">
+        <img src="../static/pic/btn_rightArrow.png" id="pic4" onmouseover="enter2()" onmouseout="leave2()" onclick="switchto2()">
         <div id="ball"></div>
     </div>
     <div class="parttwo">
@@ -53,12 +55,12 @@
             <div class="b_date" >選擇日期：<input id="date" type="date" required></div>
             <div class="b_time" >選擇時間：
                 <span onclick="changetime1()" class="halfday">
-                    <img src="../static/Ellipse 2.png" id="green1">
-                    <img src="../static/Ellipse 1.png" id="white1">
+                    <img src="../static/pic/Ellipse 2.png" id="green1">
+                    <img src="../static/pic/Ellipse 1.png" id="white1">
                 </span><span class="half_day1">上半天</span>
                 <span onclick="changetime2()" class="halfday">
-                    <img src="../static/Ellipse 2.png" id="green2" >
-                    <img src="../static/Ellipse 1.png" id="white2">
+                    <img src="../static/pic/Ellipse 2.png" id="green2" >
+                    <img src="../static/pic/Ellipse 1.png" id="white2">
                 </span><span class="half_day2">下半天</span>
             </div>
             <div class="b_fee">導覽費用：<span id="money2000">新台幣 2000 元</span><span id="money2500">新台幣 2500 元</span></div>
@@ -77,5 +79,6 @@
         <div class="copyright" id="copyright">COPYRIGHT © 2021 台北一日遊</div>
     </div>
     <script type="text/javascript" src="../static/attraction.js"></script>
+    <script type="text/javascript" src="../static/LoginoutRegister.js"></script>
 </body>
 </html>
