+++ templates/booking.html
@@ -4,9 +4,10 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/png" sizes="32x32" href="../static/bx-building-house.svg">
+    <link rel="icon" type="image/png" sizes="32x32" href="../static/pic/bx-building-house.svg">
     <title>台北一日遊</title>
     <link href="../static/booking.css" rel="stylesheet"/>
+    <link href="../static/same.css" rel="stylesheet"/>
 </head>
 <body onload="checkStatus()">
     <div class="title">
@@ -28,7 +29,7 @@
             <div class="text3">時間：<span id="b_time"></span></div>
             <div class="text4">費用：新台幣 <span id="b_fee"></span> 元</div>
             <div class="text5">地點：<span id="b_address"></span></div>
-            <img src="../static/icon_delete.png" class="delete_icon" onclick="deletebooking()">
+            <img src="../static/pic/icon_delete.png" class="delete_icon" onclick="deletebooking()">
         </div>
         </div>
     </div>
