+++ templates/index.html
@@ -4,9 +4,11 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/png" sizes="32x32" href="../static/bx-building-house.svg">
+    <link rel="icon" type="image/png" sizes="32x32" href="../static/pic/bx-building-house.svg">
     <title>台北一日遊</title>
     <link href="../static/index.css" rel="stylesheet"/> 
+    <link href="../static/same.css" rel="stylesheet"/>
+    <link href="../static/signinup.css" rel="stylesheet"/>
 </head>
 <body onload="getData(0)">
     <div id="signin">
@@ -44,10 +46,10 @@
         <div class="slogan">探索每個角落，體驗城市的深度旅遊行程</div>
         <div class="query">
         <input type="text" placeholder="輸入景點名稱查詢" id="query">
-        <button class="button" onclick="remove(query(0))"><img src="../static/icon_search.png" class="img1" /></button>
+        <button class="button" onclick="remove(query(0))"><img src="../static/pic/icon_search.png" class="img1" /></button>
         </div>
         
-        <img src="../static/welcome.png" class="picture"> 
+        <img src="../static/pic/welcome.png" class="picture"> 
     </div>
     <div id="attractions" class="attractions">
     </div>
@@ -56,5 +58,6 @@
     </div>
     
     <script type="text/javascript" src="../static/index.js"></script>
+    <script type="text/javascript" src="../static/LoginoutRegister.js"></script>
 </body>
 </html>
